backend/mlflow_system/requirements.txt
# SutazAI MLflow System Requirements
# Comprehensive experiment tracking and ML pipeline automation

# Core MLflow Components
mlflow==2.9.2
mlflow-server==2.9.2

# Database Support
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.29.0

# Redis for Messaging
redis==5.2.1
aioredis==2.0.1

# HTTP and API Support
httpx==0.28.1
aiohttp==3.11.11
requests==2.32.3
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Data Science and ML
pandas==2.2.3
numpy==2.1.3
scipy==1.14.1
scikit-learn==1.5.2

# Hyperparameter Optimization
optuna==4.1.0
bayesian-optimization==1.5.1

# Visualization and Dashboards
streamlit==1.41.1
plotly==5.24.1
matplotlib==3.9.4
seaborn==0.13.2

# ML Framework Support
torch==2.5.1
tensorflow==2.18.0
xgboost==2.1.3

# Monitoring and Metrics
prometheus-client==0.21.1
psutil==6.1.0

# Configuration and Utilities
pydantic==2.10.4
pydantic-settings==2.8.1
python-dotenv==1.0.1
pyyaml==6.0.2
click==8.1.8
rich==13.9.4
typer==0.15.1

# Async Support
asyncio-mqtt==0.16.2

# Development and Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
black==24.10.0
isort==5.13.2
mypy==1.13.0

# Security
cryptography==44.0.0
python-jose[cryptography]==3.3.0

# Additional Utilities
pathlib2==2.3.7
dataclasses-json==0.6.7
tenacity==9.0.0